parity-v2 green: Phase 3 electron DRY migrations
T-P3.1 voice-chat: deleted apps/rishi-electron/src/renderer/src/services/voice-chat
(19 files) in favor of @rishi/shared/voice-chat. Ported electron's
realtime usage accumulator + billing wires INTO shared and threaded
captureError through deps as an optional port. Electron supplies
sentry.captureError and billing: { apiFetch: workerFetch } at the
factory call site.

T-P3.2 tts: deleted 12 of 14 electron tts files. Kept electron-specific
glue (index.ts now re-exports from @rishi/shared/tts plus the no-arg
resolveParagraphElement bridging the renderer EPUB-frame registry, and
the singleton getVisualCueEmitter wrapper).

T-P3.3 book-import: SCOPED OUT (see .parity-v2/MIGRATION-NOTES.md).
Electron's and shared's book-import diverged too deeply for an
adapter — electron has scanner/discovery + hash stage + R2 fileSync
that shared dropped, and shared is generic over BookId. Defer to a
dedicated phase that ports electron's discovery + hash INTO shared.

T-P3.4 prompt helpers: deleted ~95 LoC of inline render helpers in
buildRealtimeAgent.ts; now calls renderRealtimeInstructions from
@rishi/shared/voice-chat. Byte-identical prompt output.

T-P3.5 navigationHistory: deleted 6 files; navigationHistoryActor.ts
now imports navigationHistoryMachine from
@rishi/shared/machines/navigationHistory. types.ts is a re-export
shim. No Map -> Record adapter needed (no electron consumers read
context.resumeMap directly).

T-P3.6 wiring merge: resolved both PARITY-V2 markers. Swapped 5
consumer-side imports (chatStore.ts + 4 test files) from
@/services/voice-chat -> @rishi/shared/voice-chat.

Test counts:
- @rishi/shared vitest: 51 files, 653 passed + 2 expected fail
- rishi-electron renderer vitest: 183 files, 1443 passed (0 failures)
- rishi-electron full vitest: 3 pre-existing better-sqlite3
  NODE_MODULE_VERSION failures in src/main (verified identical on main
  via git stash round-trip)
- rishi-electron tsc --noEmit: clean

LoC delta: electron -~5230 LoC, shared +~260 LoC; net repo ~-4970 LoC
while preserving electron behavior bit-for-bit.
