--- apps/rishi-electron/src/renderer/src/services/tts/cache.test.ts
@@ -1,236 +1,0 @@
-import { describe, it, expect, vi } from 'vitest'
-import type { TtsIpcChannels } from './types'
-import { createCache } from './cache'
-
-/**
- * Build a fake TtsIpcChannels backed by an in-memory file map.
- * Exposes spy access via vi.fn() so tests can assert on call sequence.
- */
-export function makeIpc(initial: Record<string, Uint8Array> = {}): {
-  ipc: TtsIpcChannels
-  files: Map<string, Uint8Array>
-  // True if `dest` was created via linkOrCopyFile and points at the same
-  // underlying bytes as `src` (a hardlink in real life). The in-memory
-  // fake models this by storing the same Uint8Array reference under both
-  // keys so tests can compare identity to assert sharing.
-  isLinked(dest: string): boolean
-} {
-  const files = new Map<string, Uint8Array>(Object.entries(initial))
-  const dirs = new Set<string>()
-  const linked = new Set<string>()
-  const ipc: TtsIpcChannels = {
-    mkdir: vi.fn(async (path) => {
-      dirs.add(path)
-    }),
-    exists: vi.fn(async (path) => files.has(path) || dirs.has(path)),
-    writeFile: vi.fn(async (path, data) => {
-      files.set(path, new Uint8Array(data))
-    }),
-    readFile: vi.fn(async (path) => {
-      const f = files.get(path)
-      if (!f) throw new Error(`ENOENT: ${path}`)
-      return f.buffer.slice(f.byteOffset, f.byteOffset + f.byteLength) as ArrayBuffer
-    }),
-    copyFile: vi.fn(async (src, dest) => {
-      const f = files.get(src)
-      if (!f) throw new Error(`ENOENT: ${src}`)
-      files.set(dest, new Uint8Array(f))
-    }),
-    linkOrCopyFile: vi.fn(async (src, dest) => {
-      const f = files.get(src)
-      if (!f) throw new Error(`ENOENT: ${src}`)
-      // Share the same Uint8Array reference under both keys to model an
-      // inode hardlink: tests can assert files.get(src) === files.get(dest)
-      // to prove no copy happened.
-      files.set(dest, f)
-      linked.add(dest)
-    }),
-    removeFile: vi.fn(async (path) => {
-      files.delete(path)
-      dirs.delete(path)
-      // recursive rm: also drop any keys under this prefix
-      for (const k of [...files.keys()]) {
-        if (k.startsWith(path + '/')) files.delete(k)
-      }
-    }),
-    getDirSize: vi.fn(async (path) => {
-      let total = 0
-      for (const [k, v] of files) {
-        if (k === path || k.startsWith(path + '/')) total += v.byteLength
-      }
-      return total
-    }),
-    getCacheFileStats: vi.fn(async (dir) => {
-      const out: Array<{ path: string; size: number; mtimeMs: number }> = []
-      let i = 0
-      for (const [k, v] of files) {
-        if (k.startsWith(dir + '/')) {
-          out.push({ path: k, size: v.byteLength, mtimeMs: 1000 + i++ })
-        }
-      }
-      return out
-    }),
-    getAppDataPath: vi.fn(async () => '/userData')
-  }
-  return { ipc, files, isLinked: (dest: string) => linked.has(dest) }
-}
-
-describe('cache.audioPath', () => {
-  it('resolves to `<appData>/tts-cache/<bookId>/<md5(cfiRange)>.mp3` and creates the book dir on first use', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-
-    const path = await cache.audioPath('book-1', 'epubcfi(/6/4!/4/2,/1:0,/1:10)')
-
-    // md5('epubcfi(/6/4!/4/2,/1:0,/1:10)') — we don't pin the literal hash; just structure.
-    expect(path).toMatch(/^\/userData\/tts-cache\/book-1\/[a-f0-9]{32}\.mp3$/)
-    expect(ipc.mkdir).toHaveBeenCalledWith('/userData/tts-cache')
-    expect(ipc.mkdir).toHaveBeenCalledWith('/userData/tts-cache/book-1')
-  })
-})
-
-describe('cache.saveAudio', () => {
-  it('writes the bytes under the CFI key and links (not copies) to the text-hash key', async () => {
-    const { ipc, files, isLinked } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    const bytes = new Uint8Array([1, 2, 3, 4])
-
-    const path = await cache.saveAudio('book-1', 'cfi-x', bytes, 'hello world')
-
-    expect(path).toMatch(/\.mp3$/)
-    expect(ipc.writeFile).toHaveBeenCalledTimes(1)
-    // Mirror is hardlinked, not copied — same inode, zero extra disk.
-    // copyFile must not run; linkOrCopyFile runs exactly once.
-    expect(ipc.copyFile).not.toHaveBeenCalled()
-    expect(ipc.linkOrCopyFile).toHaveBeenCalledTimes(1)
-    // Both keys present in the in-memory FS
-    const mp3s = [...files.keys()].filter((k) => k.endsWith('.mp3'))
-    expect(mp3s).toHaveLength(2)
-    // Mirror points at the same underlying Uint8Array (modeling shared
-    // inode) so the texthash lookup is free in disk-space terms.
-    const [primary, mirror] = mp3s
-    expect(files.get(primary)).toBe(files.get(mirror))
-    expect(isLinked(mirror)).toBe(true)
-  })
-
-  it('rejects empty audio buffers without writing', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-
-    await expect(cache.saveAudio('book-1', 'cfi-x', new Uint8Array(0))).rejects.toThrow(
-      'Audio blob is zero bytes'
-    )
-    expect(ipc.writeFile).not.toHaveBeenCalled()
-  })
-
-  it('does not duplicate the text-hash mirror when cfiRange already starts with texthash:', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    const bytes = new Uint8Array([1, 2, 3])
-
-    await cache.saveAudio('book-1', 'texthash:abc', bytes, 'hello')
-
-    expect(ipc.writeFile).toHaveBeenCalledTimes(1)
-    expect(ipc.copyFile).not.toHaveBeenCalled()
-    expect(ipc.linkOrCopyFile).not.toHaveBeenCalled()
-  })
-})
-
-describe('cache.getAudio', () => {
-  it('returns the cached ArrayBuffer when the CFI key hits', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    await cache.saveAudio('book-1', 'cfi-x', new Uint8Array([9, 9, 9]))
-
-    const got = await cache.getAudio('book-1', 'cfi-x')
-
-    expect(got).not.toBeNull()
-    expect(new Uint8Array(got!)).toEqual(new Uint8Array([9, 9, 9]))
-  })
-
-  it('falls back to text-hash key when CFI is absent', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    // Save under cfi-x WITH textHash so mirror is created
-    await cache.saveAudio('book-1', 'cfi-x', new Uint8Array([7, 7]), 'paragraph text')
-
-    // Looking up a different CFI but with the same text-hash should resolve
-    const got = await cache.getAudio('book-1', 'cfi-different', 'paragraph text')
-
-    expect(got).not.toBeNull()
-    expect(new Uint8Array(got!)).toEqual(new Uint8Array([7, 7]))
-  })
-
-  it('returns null when both lookups miss', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-
-    const got = await cache.getAudio('book-1', 'cfi-missing', 'no-text')
-
-    expect(got).toBeNull()
-  })
-
-  it('returns null when the CFI matches but the text has changed (no stale bytes)', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    // Same slot-positional CFI (e.g. page*10000+slotN) cached for old footnote text
-    await cache.saveAudio('book-1', 'N0001', new Uint8Array([1, 2, 3]), 'footnote text')
-
-    // After reading-order change, slot N0001 now holds different body text
-    const got = await cache.getAudio('book-1', 'N0001', 'different body text')
-
-    expect(got).toBeNull()
-  })
-})
-
-describe('cache.clearBook', () => {
-  it('removes the book directory recursively and forgets its membership', async () => {
-    const { ipc, files } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-    await cache.saveAudio('book-1', 'cfi-a', new Uint8Array([1]))
-    await cache.saveAudio('book-1', 'cfi-b', new Uint8Array([2]))
-    expect([...files.keys()].filter((k) => k.includes('/book-1/'))).toHaveLength(2)
-
-    await cache.clearBook('book-1')
-
-    expect([...files.keys()].filter((k) => k.includes('/book-1/'))).toHaveLength(0)
-    expect(ipc.removeFile).toHaveBeenCalledWith('/userData/tts-cache/book-1')
-  })
-
-  it('is a no-op when the book directory does not exist', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 500 * 1024 * 1024 })
-
-    await expect(cache.clearBook('never-cached')).resolves.toBeUndefined()
-    expect(ipc.removeFile).not.toHaveBeenCalled()
-  })
-})
-
-describe('cache.evictIfNeeded', () => {
-  it('removes oldest files until total size is under threshold', async () => {
-    const { ipc, files } = makeIpc()
-    // tight cap: 10 bytes total, threshold = 0.8 * 10 = 8 bytes
-    const cache = createCache({ ipc, cacheMaxBytes: 10 })
-    // Three 4-byte files = 12 bytes total → over 80% threshold
-    await cache.saveAudio('book-1', 'cfi-1', new Uint8Array([1, 1, 1, 1]))
-    await cache.saveAudio('book-1', 'cfi-2', new Uint8Array([2, 2, 2, 2]))
-    await cache.saveAudio('book-1', 'cfi-3', new Uint8Array([3, 3, 3, 3]))
-
-    await cache.evictIfNeeded()
-
-    // Should evict oldest file(s) until ≤ 8 bytes remain
-    const totalBytes = [...files.values()].reduce((n, b) => n + b.byteLength, 0)
-    expect(totalBytes).toBeLessThanOrEqual(8)
-    expect(ipc.removeFile).toHaveBeenCalled()
-  })
-
-  it('is a no-op when total size is below threshold', async () => {
-    const { ipc } = makeIpc()
-    const cache = createCache({ ipc, cacheMaxBytes: 1000 })
-    await cache.saveAudio('book-1', 'cfi-1', new Uint8Array([1, 2, 3]))
-
-    await cache.evictIfNeeded()
-
-    expect(ipc.removeFile).not.toHaveBeenCalled()
-  })
-})
