--- apps/rishi-electron/src/renderer/src/services/tts/emitter.ts
@@ -1,23 +1,0 @@
-/**
- * Tiny typed emitter — no event names, single payload type T.
- * `on(listener)` returns an unsubscribe function (idempotent).
- */
-export interface Emitter<T> {
-  emit(payload: T): void
-  on(listener: (payload: T) => void): () => void
-}
-
-export function createEmitter<T>(): Emitter<T> {
-  const listeners = new Set<(payload: T) => void>()
-  return {
-    emit(payload) {
-      for (const listener of listeners) listener(payload)
-    },
-    on(listener) {
-      listeners.add(listener)
-      return () => {
-        listeners.delete(listener)
-      }
-    }
-  }
-}
