+++ apps/rishi-electron/src/renderer/src/services/tts/index.ts
@@ -1,4 +1,10 @@
-export { createTtsService } from './service'
+/**
+ * Electron TTS glue. The TTS service implementation now lives in
+ * `@rishi/shared/tts`; this barrel re-exports the public surface plus the
+ * electron-specific helpers (`getVisualCueEmitter`, `resolveParagraphElement`)
+ * that reach into renderer-only globals (`getActiveEpubFrame` registry).
+ */
+export { createTtsService } from '@rishi/shared/tts'
 export type {
   AudioRequest,
   AudioReadyEvent,
@@ -9,9 +15,13 @@ export type {
   TtsIpcChannels,
   TtsService,
   TtsServiceDeps
-} from './types'
+} from '@rishi/shared/tts'
 
-import { createVisualCueEmitter, type VisualCueEmitter } from './visual-cue-emitter'
+import {
+  createVisualCueEmitter,
+  resolveParagraphElement as resolveParagraphElementInBody,
+  type VisualCueEmitter
+} from '@rishi/shared/tts'
 
 let _visualCueEmitter: VisualCueEmitter | null = null
 
@@ -20,7 +30,7 @@ export function getVisualCueEmitter(): VisualCueEmitter {
   return _visualCueEmitter
 }
 
-export type { VisualNearbyEvent } from './visual-cue-emitter'
+export type { VisualNearbyEvent } from '@rishi/shared/tts'
 
 import { getActiveEpubFrame } from '@/modules/pageCapture/epubFrameRegistry'
 
@@ -29,14 +39,13 @@ import { getActiveEpubFrame } from '@/modules/pageCapture/epubFrameRegistry'
  * reader. Returns null when no reasonable element can be located (PDF mode,
  * iframe not yet rendered, paragraph out of range). Used by the TTS visual
  * cue — null disables the cue silently.
+ *
+ * Thin electron-side wrapper around `@rishi/shared/tts`'s pure
+ * `resolveParagraphElement(body, index)` — supplies the body element from the
+ * renderer-only EPUB-frame registry.
  */
 export function resolveParagraphElement(paragraphIndexInPage: number): Element | null {
   const frame = getActiveEpubFrame()
   const body = frame?.contentDocument?.body
-  if (!body) return null
-  // Match common block-level elements that typically map to paragraphs in
-  // EPUB content. Order-preserving querySelectorAll guarantees positional
-  // correspondence with paragraphs published by the EPUB reader.
-  const candidates = body.querySelectorAll('p, h1, h2, h3, h4, h5, h6, blockquote, li')
-  return candidates[paragraphIndexInPage] ?? null
+  return resolveParagraphElementInBody(body, paragraphIndexInPage)
 }
