--- apps/rishi-electron/src/renderer/src/services/voice-chat/index.ts
@@ -1,31 +1,0 @@
-export { createVoiceChatService } from './service'
-export { createLocalVad, VadTimeoutError, VadDisposedError } from './local-vad'
-export { OfflineError } from './types'
-export type {
-  ChatStatus,
-  VoiceChatPublicState,
-  VoiceChatService,
-  VoiceChatServiceDeps,
-  VoiceChatContext,
-  VoiceChatConfig,
-  VoiceChatIpc,
-  VoiceError,
-  VoiceErrorReason,
-  MediaPort,
-  NetworkPort,
-  EffectsPort,
-  ClockPort,
-  AgentFactoryArgs,
-  WebrtcFactoryArgs,
-  SessionFactoryOpts,
-  RealtimeAgentLike,
-  RealtimeSessionLike,
-  RtcTransportLike,
-  MediaStreamLike,
-  MediaRecorderLike,
-  AudioElementLike,
-  LocalVadConfig,
-  LocalVoiceVad,
-  ServerVadConfig,
-  VadPort
-} from './types'
