+++ apps/rishi-electron/src/renderer/src/stores/initBookChatSubscription.test.ts
@@ -34,7 +34,7 @@ const { FakeOfflineError } = vi.hoisted(() => ({
   }
 }))
 
-vi.mock('@/services/voice-chat', () => ({
+vi.mock('@rishi/shared/voice-chat', () => ({
   OfflineError: FakeOfflineError
 }))
